Added reset counter and commands to read it out and set it back to zero
The a seperate usb_read routine was added which does not wait an infinite time
until RXF goes low. So the main loop, waiting for commands via USB always
returns from the read_usb routine to feed the watchdog.
All the other read_usb calls use the function without timeout, to prevent
communication probelmes when the values after the commands are transmited

Everythinh is working